Add conversion from prefs to pref string
Add a conversion from the prefs (taken from W2W) to the string format
detailed at the top of the file. This allows for easier reading of prefs
down the road, and is a more sensible representation than that on W2W
itself.

Add some additional documentation at the top.